Refactor several aspects of the code
Now managers are not changing the local state, and just work remotely
and only append requests to the queue (thus they are suited for
applications that want a more stateless approach).

Also, the models' instances (objects) are using the aforementioned
managers to append requests to the queue, and additionally they are
modifying themselves, too.

And by having the objects reuse the managers functions, we avoid
having similar code twice (the appending of requests to the queue part).