Improve the README
The README were too simple. This commit adds some more examples and
other improvements.

There was no examples on how to use the library to add a task, just to
get information from Todoist. This commit adds examples on how to add
and update a task, also linking to the documentation of the modules
used.